+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Change Log
 
 - Optimize the performance of the incompatible character scan.
 - [trivial] Add tooltip to font field in Appearance settings pane.
+- [dev] Migrate the custom sort pattern view to SwiftUI.
 
 
 ### Fixes